Producing output for first time
But getting a blank screen when viewing with chrome. Discovered that
inch units are not allowed inside paths. You must use so called
"pixels" which are defined to be 1/96 of an inch. Fractions are
allowed.  So I have to change the scale factor.  Not a huge deal.

There are also other problems problems with scaling.  There are some
errors in the scaling equations. It may be simpler to use absolute
moves.