Got text display in svg figured out
Title and refs and angle and order are working.
Still need to do rules.

Have working example for hyperlinks for references.
Although it does not change color when you hover over it.

Finally got a working quadratic gosper curve.
Although it gets big very quickly and it is almost too small to see
clearly in order 3,.